PramaIA-Agents/document-folder-monitor-agent/requirements.txt
watchdog
pdfplumber
sentence-transformers
qdrant-client
fastapi
uvicorn
python-dotenv
requests
sqlite3
asyncio
aiohttp
# PramaIALogger - Installazione locale dal percorso relativo
../PramaIA-LogService/clients/python/
